refactor: rename folder app to application
