+++ src/application/auth/services/signInUser/helpers/createJwtToken.go
