+++ src/infra/router/routes.go
@@ -3,8 +3,8 @@ package routes
 import (
 	"github.com/gofiber/fiber/v2"
 	"github.com/gofiber/fiber/v2/middleware/cors"
-	authController "github.com/rodrigoRVSN/beeus-api/src/app/auth/controllers"
-	userController "github.com/rodrigoRVSN/beeus-api/src/app/users/controllers"
+	authController "github.com/rodrigoRVSN/beeus-api/src/application/auth/controllers"
+	userController "github.com/rodrigoRVSN/beeus-api/src/application/users/controllers"
 )
 
 func SetupRoutes(app *fiber.App) {
